chore(components/ProductCard): refactor ProductCard subcomponents
