+++ src/components/ProductCard/Main/Main.module.scss
