--- src/components/ProductCard/ProductCardHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './ProductCardHeader';
