--- src/components/ProductCard/ProductCardMain/index.ts
@@ -1,1 +1,0 @@
-export * from './ProductCardMain';
