--- src/components/ProductCard/ProductCardThumbnail/index.ts
@@ -1,1 +1,0 @@
-export * from './ProductCardThumbnail';
