+++ src/components/ProductCard/Thumbnail/Thumbnail.module.scss
