+++ src/components/ProductCard/Thumbnail/Thumbnail.tsx
@@ -2,13 +2,13 @@ import clsx from 'clsx';
 import { Eye, Heart, ShoppingCart } from 'lucide-react';
 import { Product } from '../../../types/products';
 import { Card } from '../../Card';
-import hoverStyles from '../ProductCard.module.scss';
-import styles from './ProductCardThumbnail.module.scss';
 import { ProductQuickViewModal } from '../../ProductQuickViewModal';
+import hoverStyles from '../ProductCard.module.scss';
+import styles from './Thumbnail.module.scss';
 
-type ProductCardThumbnailProps = { product: Product };
+type ThumbnailProps = { product: Product };
 
-export const ProductCardThumbnail = ({ product }: ProductCardThumbnailProps) => (
+export const Thumbnail = ({ product }: ThumbnailProps) => (
   <Card.Thumbnail
     src={product.thumbnail}
     alt={`${product.title} photo`}
