+++ src/components/ProductCard/Thumbnail/index.ts
@@ -1,0 +1,1 @@
+export * from './Thumbnail';
